https://gitee.com/openharmony/arkui_ace_engine/issues/IABCKQ Description:修复告警 Sig: SIG_ApplicationFramework Feature or Bugfix:Bugfix Binary Source:No TDD:(评估不涉及) XTS:Pass 预测试:评估不涉及 Signed-off-by:guanzengkun <guanzengkun@huawei.com>

+++ frameworks/core/components_ng/pattern/text/text_pattern.cpp
@@ -1373,9 +1373,9 @@ double TextPattern::GetTextHeight(int32_t index, bool isNextLine)
         } else {
             auto textLayoutProperty = GetLayoutProperty<TextLayoutProperty>();
             CHECK_NULL_RETURN(textLayoutProperty, 0);
-            auto maxLines = textLayoutProperty->GetMaxLinesValue(Infinity<float>());
+            auto maxLines = textLayoutProperty->GetMaxLinesValue(Infinity<uint32_t>());
             if ((index <= endIndex && startIndex != 0) ||
-                ((lineNumber + 1) == maxLines && lineNumber != 0)) {
+                ((lineNumber + 1) == static_cast<int32_t>(maxLines) && lineNumber != 0)) {
                 return GetLineMetrics(lineNumber - 1).height;
             }
         }
